Fix passage of 64-bit addresses to disassembler modules

+++ capstone.cpp
@@ -53,7 +53,7 @@ cs_mode mode = (cs_mode)(CS_MODE_64);
 cs_arch arch = (cs_arch)CS_ARCH_X86;
 #endif
 
-extern "C" int disassemble(uint32_t addr_, uint8_t *data, int len, char *result)
+extern "C" int disassemble(uint64_t addr_, uint8_t *data, int len, char *result)
 {
 	int rc = -1;
